+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @party-kit/cli
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @party-kit/core@0.1.0
+
 ## 0.0.4
 
 ### Patch Changes